pyproject.toml
[tool.poetry]
name = "habitcreation"
version = "0.1.0"
description = ""
authors = ["Renttor <denis30510@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.6"
celery = "^5.3.4"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.0"
djangorestframework = "^3.14.0"
ipython = "^8.16.1"
drf-yasg = "^1.21.7"
redis = "^5.0.1"
django-celery-beat = "^2.5.0"
coverage = "^7.3.2"
stripe = "^7.0.0"
djangorestframework-simplejwt = "^5.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
